// pre-deployment/types.ts
export interface PlotSeriesData {
  [key: string]: string | number;
}

export interface PlotLineConfig {
    key: string;
    name: string;
    color: string;
}

export interface PlotThresholdLine {
    label: string;
    value: number;
    color: string;
}

export interface PlotOptions {
    xlabel: string;
    ylabel: string;
    stacked?: boolean;
    invert_y?: boolean;
    threshold_lines?: PlotThresholdLine[];
    layout?: 'horizontal' | 'vertical';
}

export interface PlotConfig {
    id: string;
    type: 'bar' | 'line';
    title: string;
    x_field: string;
    y_fields: PlotLineConfig[];
    series: PlotSeriesData[];
    options: PlotOptions;
}